+++ libraries/launcher/org/prismlauncher/fix/Fixes.java
@@ -35,29 +35,14 @@
 
 package org.prismlauncher.fix;
 
-import java.util.Collections;
-import java.util.List;
-
-import org.prismlauncher.fix.skins.SkinFix;
+import org.prismlauncher.fix.online.OnlineFixes;
 import org.prismlauncher.utils.Parameters;
-import org.prismlauncher.utils.logging.Log;
 
 public final class Fixes {
 
-    private static final Fix[] FIXES = { new SkinFix() };
-
     public static void apply(Parameters params) {
-        List<String> fixes = params.getList("fixes", Collections.<String>emptyList());
-
-		for (Fix fix : FIXES) {
-			if (fixes.contains(fix.getName()) && fix.isApplicable(params)) {
-				try {
-					fix.apply();
-				} catch (Throwable e) {
-					Log.error("Could not apply " + fix.getName(), e);
-				}
-			}
-		}
+        if ("true".equalsIgnoreCase(params.getString("onlineFixes", null)))
+            OnlineFixes.apply();
     }
 
 }
