+++ libraries/launcher/org/prismlauncher/fix/online/SkinFix.java
@@ -1,56 +1,17 @@
-// SPDX-License-Identifier: GPL-3.0-only
-/*
- *  Prism Launcher - Minecraft Launcher
- *  Copyright (C) 2022 TheKodeToad <TheKodeToad@proton.me>
- *
- *  This program is free software: you can redistribute it and/or modify
- *  it under the terms of the GNU General Public License as published by
- *  the Free Software Foundation, version 3.
- *
- *  This program is distributed in the hope that it will be useful,
- *  but WITHOUT ANY WARRANTY; without even the implied warranty of
- *  MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
- *  GNU General Public License for more details.
- *
- *  Linking this library statically or dynamically with other modules is
- *  making a combined work based on this library. Thus, the terms and
- *  conditions of the GNU General Public License cover the whole
- *  combination.
- *
- *  As a special exception, the copyright holders of this library give
- *  you permission to link this library with independent modules to
- *  produce an executable, regardless of the license terms of these
- *  independent modules, and to copy and distribute the resulting
- *  executable under terms of your choice, provided that you also meet,
- *  for each linked independent module, the terms and conditions of the
- *  license of that module. An independent module is a module which is
- *  not derived from or based on this library. If you modify this
- *  library, you may extend this exception to your version of the
- *  library, but you are not obliged to do so. If you do not wish to do
- *  so, delete this exception statement from your version.
- *
- *  You should have received a copy of the GNU General Public License
- *  along with this program.  If not, see <https://www.gnu.org/licenses/>.
- */
-
-package org.prismlauncher.fix.skins;
+package org.prismlauncher.fix.online;
 
 import java.io.IOException;
 import java.io.InputStream;
-import java.net.Proxy;
 import java.net.URL;
-import java.net.URLConnection;
-import java.net.URLStreamHandler;
 import java.util.Map;
 
 import org.prismlauncher.utils.Base64;
 import org.prismlauncher.utils.JsonParser;
-import org.prismlauncher.utils.UrlUtils;
 
 @SuppressWarnings("unchecked")
-final class Handler extends URLStreamHandler {
+final class SkinFix {
 
-    private URL redirect(URL address) throws IOException {
+    static URL redirect(URL address) throws IOException {
         String skinOwner = findSkinOwner(address);
         if (skinOwner != null)
             return convert(skinOwner, "SKIN");
@@ -62,18 +23,7 @@ final class Handler extends URLStreamHandler {
         return address;
     }
 
-    @Override
-    protected URLConnection openConnection(URL address) throws IOException {
-        return openConnection(address, null);
-    }
-
-    @Override
-    protected URLConnection openConnection(URL address, Proxy proxy) throws IOException {
-        address = redirect(address);
-        return UrlUtils.openHttpConnection(address, proxy);
-    }
-
-    private URL convert(String owner, String name) throws IOException {
+    private static URL convert(String owner, String name) throws IOException {
         Map<String, Object> textures = getTextures(owner);
 
         if (textures != null) {
